+++ tests/testthat/test-query-set-op.R
@@ -9,4 +9,8 @@ test_that("generated sql doesn't change unexpectedly", {
   expect_snapshot(union(lf, lf))
   expect_snapshot(setdiff(lf, lf))
   expect_snapshot(intersect(lf, lf))
+
+  expect_snapshot(union(lf, lf, all = TRUE))
+  expect_snapshot(setdiff(lf, lf, all = TRUE))
+  expect_snapshot(intersect(lf, lf, all = TRUE))
 })